# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MCQ Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mcqs:
    get:
      operationId: getMcqs
      tags: [mcqs]
      summary: List MCQs with optional filters
      parameters:
        - name: chapterId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, bookmarked, wrong]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of MCQs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McqListResponse"
    post:
      operationId: createMcq
      tags: [mcqs]
      summary: Create an MCQ (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McqInput"
      responses:
        "201":
          description: MCQ created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mcq"
  /mcqs/bulk:
    post:
      operationId: bulkCreateMcqs
      tags: [mcqs]
      summary: Bulk create MCQs (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkMcqInput"
      responses:
        "201":
          description: MCQs created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkMcqResponse"
  /mcqs/{id}:
    get:
      operationId: getMcq
      tags: [mcqs]
      summary: Get an MCQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: MCQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mcq"
    patch:
      operationId: updateMcq
      tags: [mcqs]
      summary: Update an MCQ (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McqUpdate"
      responses:
        "200":
          description: MCQ updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mcq"
    delete:
      operationId: deleteMcq
      tags: [mcqs]
      summary: Delete an MCQ (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /mcqs/{id}/submit:
    post:
      operationId: submitAnswer
      tags: [mcqs]
      summary: Submit an answer for an MCQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"

  # Bookmarks
...
  /exams:
    get:
      operationId: getExams
      tags: [exams]
      summary: List exam sessions for current user
      responses:
        "200":
          description: List of exam sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamSession"
    post:
      operationId: startExam
      tags: [exams]
      summary: Start a new exam session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamStartInput"
      responses:
        "201":
          description: Exam started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamSession"
  /exams/{id}:
    get:
      operationId: getExam
      tags: [exams]
      summary: Get exam session details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamSession"
  /exams/{id}/submit:
    post:
      operationId: submitExam
      tags: [exams]
      summary: Submit exam answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamSubmitInput"
      responses:
        "200":
          description: Exam result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"

  # Results
...
  /results/{id}:
    get:
      operationId: getResult
      tags: [results]
      summary: Get a specific result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"

  # Stats
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth schemas
...
    Mcq:
      type: object
      required: [id, chapterId, subjectId, question, options, correctOption, difficulty, isBookmarked, isWrong]
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        subjectId:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: ["string", "null"]
        explanationImage:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        isBookmarked:
          type: boolean
        isWrong:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    McqInput:
      type: object
      required: [chapterId, question, options, correctOption]
      properties:
        chapterId:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: string
        explanationImage:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: array
          items:
            type: string
    McqUpdate:
      type: object
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: string
        explanationImage:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: array
          items:
            type: string
    McqListResponse:
      type: object
      required: [mcqs, total]
      properties:
        mcqs:
          type: array
          items:
            $ref: "#/components/schemas/Mcq"
        total:
          type: integer
    BulkMcqInput:
      type: object
      required: [chapterId, mcqs]
      properties:
        chapterId:
          type: integer
        mcqs:
          type: array
          items:
            type: object
            required: [question, options, correctOption]
            properties:
              question:
                type: string
              options:
                type: array
                items:
                  type: string
              correctOption:
                type: integer
              explanation:
                type: string
              difficulty:
                type: string
    BulkMcqResponse:
      type: object
      required: [created, total]
      properties:
        created:
          type: integer
        total:
          type: integer
        mcqs:
          type: array
          items:
            $ref: "#/components/schemas/Mcq"
    AnswerInput:
      type: object
      required: [selectedOption]
      properties:
        selectedOption:
          type: integer
        examSessionId:
          type: ["integer", "null"]
    AnswerResult:
      type: object
      required: [isCorrect, correctOption, explanation]
      properties:
        isCorrect:
          type: boolean
        correctOption:
          type: integer
        explanation:
          type: ["string", "null"]
        explanationImage:
          type: ["string", "null"]

    # Exam schemas
    ExamStartInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [practice, timed, full, custom, mixed]
        subjectId:
          type: ["integer", "null"]
        chapterIds:
          type: array
          items:
            type: integer
        questionCount:
          type: integer
        timeLimitMinutes:
          type: ["integer", "null"]
    ExamSession:
      type: object
      required: [id, userId, type, status, mcqs, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        status:
          type: string
          enum: [active, completed, expired]
        timeLimitMinutes:
          type: ["integer", "null"]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        mcqs:
          type: array
          items:
            $ref: "#/components/schemas/Mcq"
        createdAt:
          type: string
    ExamSubmitInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [mcqId, selectedOption]
            properties:
              mcqId:
                type: integer
              selectedOption:
                type: integer
    ExamResult:
      type: object
      required: [id, examSessionId, userId, score, total, accuracy, timeTakenSeconds, chapterBreakdown, createdAt]
      properties:
        id:
          type: integer
        examSessionId:
          type: integer
        userId:
          type: integer
        score:
          type: integer
        total:
          type: integer
        accuracy:
          type: number
        timeTakenSeconds:
          type: integer
        chapterBreakdown:
          type: array
          items:
            type: object
            required: [chapterId, chapterName, correct, total]
            properties:
              chapterId:
                type: integer
              chapterName:
                type: string
              correct:
                type: integer
              total:
                type: integer
        createdAt:
          type: string

    # Stats schemas
...